requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
python-dotenv==1.0.1
SQLAlchemy==2.0.30        # Dependência do Flask-SQLAlchemy
Werkzeug==3.0.2           # Dependência do Flask
Click==8.1.7              # Dependência do Flask
itsdangerous==2.1.2       # Dependência do Flask
Jinja2==3.1.4             # Dependência do Flask
